requirements.txt
fastapi
uvicorn
python-dotenv
requests
python-multipart
lark-oapi
mcp
playwright
openai